third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-013-expected.txt


FAIL CSS Test: @viewport constrained - max-zoom affecting width. assert_equals: expected 1600 but got 800
